pyproject.toml
[project]
name = "twilio-manager-mcp"
version = "0.1.6"
description = "Twilio Manager MCP connects Claude AI or any host to Twilio through the Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Erraj Badr", email = "errajibadr@gmail.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Telephony",
    "Topic :: Utilities",
]
dependencies = [
    "mcp[cli]>=1.4.1",
    "twilio>=9.5.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
]

[project.urls]
"Homepage" = "https://github.com/errajibadr/twilio-manager-mcp"
"Bug Tracker" = "https://github.com/errajibadr/twilio-manager-mcp/issues"

[project.scripts]
twilio-manager-mcp = "twilio_manager_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "."]
exclude = [
    "*.env",
    ".env",
    "*.pyc",
    "__pycache__",
    "*.egg-info",
    "dist",
    "build"
]

[dependency-groups]
dev = [
]

